+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sensors/camera/utils.py
@@ -9,10 +9,13 @@ from __future__ import annotations
 import math
 import numpy as np
 import torch
+import torch.nn.functional as F
 from typing import Sequence
 from typing_extensions import Literal
 
+import omni.isaac.core.utils.stage as stage_utils
 import warp as wp
+from pxr import UsdGeom
 
 import omni.isaac.orbit.utils.math as math_utils
 from omni.isaac.orbit.utils.array import TensorData, convert_to_torch
@@ -22,6 +25,7 @@ __all__ = [
     "create_pointcloud_from_depth",
     "create_pointcloud_from_rgbd",
     "convert_orientation_convention",
+    "create_rotation_matrix_from_view",
 ]
 
 
@@ -292,9 +296,9 @@ def convert_orientation_convention(
 
     Possible conventions are:
 
-    - :obj:"opengl" - forward axis: -Z - up axis +Y - Offset is applied in the OpenGL (Usd.Camera) convention
-    - :obj:"ros"    - forward axis: +Z - up axis -Y - Offset is applied in the ROS convention
-    - :obj:"world"  - forward axis: +X - up axis +Z - Offset is applied in the World Frame convention
+    - :obj:`"opengl"` - forward axis: -Z - up axis +Y - Offset is applied in the OpenGL (Usd.Camera) convention
+    - :obj:`"ros"`    - forward axis: +Z - up axis -Y - Offset is applied in the ROS convention
+    - :obj:`"world"`  - forward axis: +X - up axis +Z - Offset is applied in the World Frame convention
 
     Args:
         orientation: Quaternion of form `(w, x, y, z)` with shape (..., 4) in source convention
@@ -348,3 +352,53 @@ def convert_orientation_convention(
         return math_utils.quat_from_matrix(rotm)
     else:
         return quat_gl.clone()
+
+
+# @torch.jit.script
+def create_rotation_matrix_from_view(
+    eyes: torch.Tensor,
+    targets: torch.Tensor,
+    device: str = "cpu",
+) -> torch.Tensor:
+    """
+    This function takes a vector ''eyes'' which specifies the location
+    of the camera in world coordinates and the vector ''targets'' which
+    indicate the position of the object.
+    The output is a rotation matrix representing the transformation
+    from world coordinates -> view coordinates.
+
+        The inputs camera_position and targets can each be a
+        - 3 element tuple/list
+        - torch tensor of shape (1, 3)
+        - torch tensor of shape (N, 3)
+
+    Args:
+        eyes: position of the camera in world coordinates
+        targets: position of the object in world coordinates
+
+    The vectors are broadcast against each other so they all have shape (N, 3).
+
+    Returns:
+        R: (N, 3, 3) batched rotation matrices
+
+    Reference:
+    Based on PyTorch3D (https://github.com/facebookresearch/pytorch3d/blob/eaf0709d6af0025fe94d1ee7cec454bc3054826a/pytorch3d/renderer/cameras.py#L1635-L1685)
+    """
+    up_axis_token = stage_utils.get_stage_up_axis()
+    if up_axis_token == UsdGeom.Tokens.y:
+        up_axis = torch.tensor((0, 1, 0), device=device, dtype=torch.float32).repeat(eyes.shape[0], 1)
+    elif up_axis_token == UsdGeom.Tokens.z:
+        up_axis = torch.tensor((0, 0, 1), device=device, dtype=torch.float32).repeat(eyes.shape[0], 1)
+    else:
+        raise ValueError(f"Invalid up axis: {up_axis_token}")
+
+    # get rotation matrix in opengl format (-Z forward, +Y up)
+    z_axis = -F.normalize(targets - eyes, eps=1e-5)
+    x_axis = F.normalize(torch.cross(up_axis, z_axis, dim=1), eps=1e-5)
+    y_axis = F.normalize(torch.cross(z_axis, x_axis, dim=1), eps=1e-5)
+    is_close = torch.isclose(x_axis, torch.tensor(0.0), atol=5e-3).all(dim=1, keepdim=True)
+    if is_close.any():
+        replacement = F.normalize(torch.cross(y_axis, z_axis, dim=1), eps=1e-5)
+        x_axis = torch.where(is_close, replacement, x_axis)
+    R = torch.cat((x_axis[:, None, :], y_axis[:, None, :], z_axis[:, None, :]), dim=1)
+    return R.transpose(1, 2)