+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sensors/ray_caster/__init__.py
@@ -11,14 +11,19 @@ from __future__ import annotations
 
 from . import patterns
 from .ray_caster import RayCaster
+from .ray_caster_camera import RayCasterCamera
+from .ray_caster_camera_cfg import RayCasterCameraCfg
 from .ray_caster_cfg import RayCasterCfg
 from .ray_caster_data import RayCasterData
 
 __all__ = [
-    # sensor
+    # ray caster
     "RayCaster",
     "RayCasterData",
     "RayCasterCfg",
+    # camera
+    "RayCasterCameraCfg",
+    "RayCasterCamera",
     # patterns
     "patterns",
 ]