+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/utils/warp/__init__.py
@@ -3,6 +3,7 @@
 #
 # SPDX-License-Identifier: BSD-3-Clause
 
+from __future__ import annotations
 
 """Operations based on warp."""